backend/requirements.txt
fastapi
uvicorn
sqlalchemy
redis
pydantic
python-dotenv
mysql-connector-python
pydantic[email]